Add BookmarkContext files
