+++ app/(features)/page/__tests__/PagePage.test.tsx
@@ -1,5 +1,6 @@
 import { render, act } from '@testing-library/react';
 import { SettingsProvider } from '@/app/providers/SettingsContext';
+import { BookmarkProvider } from '@/app/providers/BookmarkContext';
 import { AudioProvider } from '@/app/(features)/player/context/AudioContext';
 import { SidebarProvider } from '@/app/providers/SidebarContext';
 import { ThemeProvider } from '@/app/providers/ThemeContext';
@@ -65,11 +66,13 @@ const renderPage = () =>
   render(
     <AudioProvider>
       <SettingsProvider>
-        <ThemeProvider>
-          <SidebarProvider>
-            <QuranPage params={{ pageId: '1' } as unknown as Promise<{ pageId: string }>} />
-          </SidebarProvider>
-        </ThemeProvider>
+        <BookmarkProvider>
+          <ThemeProvider>
+            <SidebarProvider>
+              <QuranPage params={{ pageId: '1' } as unknown as Promise<{ pageId: string }>} />
+            </SidebarProvider>
+          </ThemeProvider>
+        </BookmarkProvider>
       </SettingsProvider>
     </AudioProvider>
   );
