+++ app/(features)/surah/__tests__/Verse.test.tsx
@@ -1,6 +1,7 @@
 import { render, screen, waitFor } from '@testing-library/react';
 import { Verse as VerseComponent } from '@/app/(features)/surah/[surahId]/components/Verse';
 import { SettingsProvider } from '@/app/providers/SettingsContext';
+import { BookmarkProvider } from '@/app/providers/BookmarkContext';
 import { AudioProvider } from '@/app/(features)/player/context/AudioContext';
 import TranslationProvider from '@/app/providers/TranslationProvider';
 import { Verse } from '@/types';
@@ -24,7 +25,9 @@ const renderVerse = () =>
     <TranslationProvider>
       <AudioProvider>
         <SettingsProvider>
-          <VerseComponent verse={verse} />
+          <BookmarkProvider>
+            <VerseComponent verse={verse} />
+          </BookmarkProvider>
         </SettingsProvider>
       </AudioProvider>
     </TranslationProvider>
