+++ app/providers/__tests__/SettingsContext.test.tsx
@@ -2,16 +2,6 @@ import { render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { SettingsProvider, useSettings } from '@/app/providers/SettingsContext';
 
-const BookmarkTest = () => {
-  const { bookmarkedVerses, toggleBookmark } = useSettings();
-  return (
-    <div>
-      <div data-testid="bookmarks">{JSON.stringify(bookmarkedVerses)}</div>
-      <button onClick={() => toggleBookmark('1:1')}>Toggle</button>
-    </div>
-  );
-};
-
 const SettingsTest = () => {
   const { settings, setSettings } = useSettings();
   return (
@@ -26,47 +16,6 @@ const SettingsTest = () => {
     </div>
   );
 };
-
-describe('SettingsContext bookmarks', () => {
-  beforeEach(() => {
-    localStorage.clear();
-  });
-
-  it('adds a verse ID to bookmarks via toggleBookmark', async () => {
-    render(
-      <SettingsProvider>
-        <BookmarkTest />
-      </SettingsProvider>
-    );
-    await userEvent.click(screen.getByRole('button'));
-    await waitFor(() => {
-      expect(screen.getByTestId('bookmarks').textContent).toBe(JSON.stringify(['1:1']));
-      expect(JSON.parse(localStorage.getItem('quranAppBookmarks') || '[]')).toEqual(['1:1']);
-    });
-  });
-
-  it('persists bookmarks in localStorage across renders', async () => {
-    const { unmount } = render(
-      <SettingsProvider>
-        <BookmarkTest />
-      </SettingsProvider>
-    );
-    await userEvent.click(screen.getByRole('button'));
-    await waitFor(() => {
-      expect(JSON.parse(localStorage.getItem('quranAppBookmarks') || '[]')).toEqual(['1:1']);
-    });
-    unmount();
-
-    render(
-      <SettingsProvider>
-        <BookmarkTest />
-      </SettingsProvider>
-    );
-    await waitFor(() => {
-      expect(screen.getByTestId('bookmarks').textContent).toBe(JSON.stringify(['1:1']));
-    });
-  });
-});
 
 describe('SettingsContext settings state', () => {
   beforeEach(() => {
